+++ inst/shiny/ui.R
@@ -390,6 +390,8 @@ fluidPage(
                     value="Median (points)\\n95% CI (horizontal lines)"),
           textInput("customcolourtitletext", label ="Pointinterval Legend title",
                     value=""),
+          textInput("customshapetitletext", label ="Shape Legend title",
+                    value=""),
           textInput("custombsvtitle", label ="BSV Legend text",
                     value="BSV (points)\\nPrediction Intervals (horizontal lines)"),
           textInput("customlinetypetitle", label ="Ref Legend text",