+++ NEWS.md
@@ -3,13 +3,13 @@
 * added citation and start up message
 * added `major_x_labels` argument for custom x labels
 * added `interval_linewidth` argument to control interval linewidth
-* added possibility to add different ref line(s) by parameter via
- `ref_value_by_panel` and `ref_value_by_panel_data` function arguments (not in the shiny, app)
 * added and exported `expand_modelframe` function
 * added the possibility to select the shapes manually via `interval_shape` and `bsv_shape`
 * added capability for user to reverse color legend separately via `legend_color_reverse`
 * added capability for user to specify text for legend titles via `interval_legend_title` and `shape_legend_title` as well as to control text size via `legend_title_size`
 * added capability for user to specify facet text color via `x_facet_text_col` and `y_facet_text_col`
+* added possibility to add different ref line(s)/area(s) by parameter via
+ `ref_value_by_panel` and `ref_value_by_panel_data` function arguments (not in the shiny, app)
 
 # coveffectsplot 1.0.3
 * fixed a bug when ref_value when not equal 1 affecting y axis limits