+++ NEWS.md
@@ -6,8 +6,8 @@
 * added and exported `expand_modelframe` function
 * added the possibility to select the shapes manually via `interval_shape` and `bsv_shape`
 * added capability for user to reverse color legend separately via `legend_color_reverse`
-* added capability for user to specify text for legend titles via `interval_legend_title` and `shape_legend_title` as well as to control text size via `legend_title_size`
-* added capability for user to specify facet text color via `x_facet_text_col` and `y_facet_text_col`
+* added capability for user to specify text for legend titles via `interval_legend_title` and `shape_legend_title` as well as to control legend title text size via `legend_title_size`
+* added capability for user to specify facet text color via `x_facet_text_col` and `y_facet_text_col` for improved theming support
 * added possibility to add different ref line(s)/area(s) by parameter via
  `ref_value_by_panel` and `ref_value_by_panel_data` function arguments (not in the shiny, app)